backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
supabase==2.0.3
redis==5.0.1
pydantic==2.5.2
pydantic-ai==0.1.0
mem0==0.1.0
archon-mcp==0.1.0
pytest==7.4.3
pytest-flask==1.3.0
gunicorn==21.2.0
requests==2.31.0
